requirements.txt
asgiref==3.5.2
cycler==0.11.0
Django==4.0.6
django-froala-editor==4.0.13
fonttools==4.34.4
gunicorn==20.1.0
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
six==1.16.0
slugify==0.0.1
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
